move timeseries static method to utils
